[Messenger] Separate unit tests from integration tests
